Refactoring   - Fixed the code according to the video
