+++ src/main/java/com/employeebook/web/controller/EmployeeController.java
@@ -1,0 +1,42 @@
+package com.employeebook.web.controller;
+
+import com.employeebook.web.exception.EmployeeAlreadyAddedException;
+import com.employeebook.web.exception.EmployeeStorageIsFullException;
+import com.employeebook.web.model.Employee;
+import com.employeebook.web.service.EmployeeService;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Collection;
+
+@RestController
+@RequestMapping("/employee")
+public class EmployeeController {
+    private final EmployeeService employeeService;
+
+    public EmployeeController(EmployeeService employeeService) {
+        this.employeeService = employeeService;
+    }
+
+    @GetMapping
+    public Collection<Employee> findAllEmployees() {
+        return employeeService.findAll();
+    }
+
+    @GetMapping(path = "/add")
+    public Employee addEmployeeCtrl(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
+        return employeeService.addEmployee(firstName, lastName);
+    }
+
+    @GetMapping(path = "/remove")
+    public Employee removeEmployeeCtrl(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
+        return employeeService.removeEmployee(firstName, lastName);
+    }
+
+    @GetMapping(path = "/find")
+    public Employee findEmployeeCtrl(@RequestParam("firstName") String firstName, @RequestParam("lastName") String lastName) {
+        return employeeService.findEmployee(firstName, lastName);
+    }
+}
