+++ src/main/java/com/employeebook/web/exception/EmployeeNotFoundException.java
@@ -1,4 +1,4 @@
-package com.employeebook.web;
+package com.employeebook.web.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
