+++ src/main/java/com/employeebook/web/model/Employee.java
@@ -1,4 +1,4 @@
-package com.employeebook.web;
+package com.employeebook.web.model;
 
 import java.util.Objects;
